Rename WebVTT polyfill
